chore: move fetching in data-table-users
