+++ apps/web/src/app/(dashboard)/admin/users/page.tsx
@@ -1,6 +1,4 @@
-import { findUsers } from '@documenso/lib/server-only/user/get-all-users';
-
-import { Users } from './users';
+import { UsersDataTable } from './data-table-users';
 
 type AdminManageUsersProps = {
   searchParams?: {
@@ -13,18 +11,10 @@ export default function AdminManageUsers({ searchParams = {} }: AdminManageUsers
   const page = Number(searchParams.page) || 1;
   const perPage = Number(searchParams.perPage) || 10;
 
-  async function search(search: string) {
-    'use server';
-
-    const results = await findUsers({ username: search, email: search, page, perPage });
-
-    return results;
-  }
-
   return (
     <div>
       <h2 className="text-4xl font-semibold">Manage users</h2>
-      <Users search={search} page={page} perPage={perPage} />
+      <UsersDataTable page={page} perPage={perPage} />
     </div>
   );
 }
